pyproject.toml
[project]
name = "framevision"
version = "0.1.0"
description = "Official code for the FRAME paper (CVPR 2025)"
readme = "README.md"
authors = [
    { name = "Andrea Boscolo Camiletto", email = "abcamiletto@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.20.2",
    "humanize>=4.12.1",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.3.0",
    "lightning>=2.5.0.post0",
    "opencv-python-headless>=4.11.0.86",
    "plotly>=6.0.1",
    "rich>=13.9.4",
    "tensorboardx>=2.6.2.2",
    "timm>=1.0.15",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "typer>=0.15.2",
    "wandb>=0.19.8",
]

[project.optional-dependencies]
all = [
    "playwright>=1.51.0",
    "beautifulsoup4>=4.13.3",
]

[project.scripts]
frame = "framevision.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 140
lint.ignore = ["E741", "E743", "F722"] # ignore ambiguous variable names
lint.extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"] # exclude imports from __init__.py

[tool.uv.sources]
torch = [{ index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }]
torchvision = [{ index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

